+++ src/components/bottomSheet/BottomSheet.tsx
@@ -19,6 +19,8 @@ import Animated, {
   runOnUI,
   cancelAnimation,
   useWorkletCallback,
+  WithSpringConfig,
+  WithTimingConfig,
 } from 'react-native-reanimated';
 import { State } from 'react-native-gesture-handler';
 import {
@@ -158,7 +160,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
       backdropComponent,
       backgroundComponent,
       footerComponent,
-      children,
+      children: Content,
 
       // accessibility
       accessible: _providedAccessible = DEFAULT_ACCESSIBLE,
@@ -311,7 +313,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     /**
      * Returns keyboard height that in the root container.
      */
-    const getKeyboardHeightInContainer = useWorkletCallback(() => {
+    const animatedKeyboardHeightInContainer = useDerivedValue(() => {
       /**
        * if android software input mode is not `adjustPan`, than keyboard
        * height will be 0 all the time.
@@ -331,7 +333,13 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
         : Math.abs(
             animatedKeyboardHeight.value - animatedContainerOffset.value.bottom
           );
-    }, [$modal, bottomInset]);
+    }, [
+      $modal,
+      android_keyboardInputMode,
+      bottomInset,
+      animatedKeyboardHeight,
+      animatedContainerOffset,
+    ]);
     //#endregion
 
     //#region state/dynamic variables
@@ -353,7 +361,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
 
       // extended position with keyboard =
       // container height - (sheet height + keyboard height in root container)
-      const keyboardHeightInContainer = getKeyboardHeightInContainer();
+      const keyboardHeightInContainer = animatedKeyboardHeightInContainer.value;
       const extendedPositionWithKeyboard = Math.max(
         0,
         animatedContainerHeight.value -
@@ -380,7 +388,15 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
       }
 
       return SHEET_STATE.OPENED;
-    }, [keyboardBehavior]);
+    }, [
+      animatedClosedPosition,
+      animatedContainerHeight,
+      animatedKeyboardHeightInContainer,
+      animatedPosition,
+      animatedSheetHeight,
+      isInTemporaryPosition,
+      keyboardBehavior,
+    ]);
     const animatedScrollableState = useDerivedValue(() => {
       /**
        * if scrollable override state is set, then we just return its value.
@@ -420,7 +436,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     });
     // dynamic
     const animatedContentHeight = useDerivedValue(() => {
-      const keyboardHeightInContainer = getKeyboardHeightInContainer();
+      const keyboardHeightInContainer = animatedKeyboardHeightInContainer.value;
       const handleHeight = Math.max(0, animatedHandleHeight.value);
       let contentHeight = animatedSheetHeight.value - handleHeight;
 
@@ -475,7 +491,15 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
        * @link (https://github.com/gorhom/react-native-bottom-sheet/issues/470)
        */
       return Math.max(contentHeight, 0);
-    }, [keyboardBehavior]);
+    }, [
+      animatedContainerHeight,
+      animatedHandleHeight,
+      animatedKeyboardHeightInContainer,
+      animatedKeyboardState,
+      animatedSheetHeight,
+      isInTemporaryPosition,
+      keyboardBehavior,
+    ]);
     const animatedIndex = useDerivedValue(() => {
       const adjustedSnapPoints = animatedSnapPoints.value.slice().reverse();
       const adjustedSnapPointsIndexes = animatedSnapPoints.value
@@ -581,7 +605,8 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
           keyboardState === KEYBOARD_STATE.SHOWN
         ) {
           isInTemporaryPosition.value = true;
-          const keyboardHeightInContainer = getKeyboardHeightInContainer();
+          const keyboardHeightInContainer =
+            animatedKeyboardHeightInContainer.value;
           return Math.max(0, highestSnapPoint - keyboardHeightInContainer);
         }
 
@@ -591,7 +616,19 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
 
         return snapPoints[currentIndex];
       },
-      [keyboardBehavior, keyboardBlurBehavior]
+      [
+        animatedContentGestureState,
+        animatedCurrentIndex,
+        animatedHandleGestureState,
+        animatedHighestSnapPoint,
+        animatedKeyboardHeightInContainer,
+        animatedKeyboardState,
+        animatedPosition,
+        animatedSnapPoints,
+        isInTemporaryPosition,
+        keyboardBehavior,
+        keyboardBlurBehavior,
+      ]
     );
     const handleOnChange = useCallback(
       function handleOnChange(index: number) {
@@ -678,7 +715,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
       animatedAnimationState.value = ANIMATION_STATE.STOPPED;
     }, [animatedPosition, animatedAnimationState, animatedAnimationSource]);
     const animateToPositionCompleted = useWorkletCallback(
-      function animateToPositionCompleted(isFinished: boolean) {
+      function animateToPositionCompleted(isFinished?: boolean) {
         isForcedClosing.value = false;
 
         if (!isFinished) {
@@ -705,7 +742,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
         position: number,
         source: ANIMATION_SOURCE,
         velocity: number = 0,
-        configs?: Animated.WithTimingConfig | Animated.WithSpringConfig
+        configs?: WithTimingConfig | WithSpringConfig
       ) {
         if (
           position === animatedPosition.value ||
@@ -776,7 +813,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     const handleSnapToIndex = useCallback(
       function handleSnapToIndex(
         index: number,
-        animationConfigs?: Animated.WithSpringConfig | Animated.WithTimingConfig
+        animationConfigs?: WithSpringConfig | WithTimingConfig
       ) {
         const snapPoints = animatedSnapPoints.value;
         invariant(
@@ -835,7 +872,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     const handleSnapToPosition = useWorkletCallback(
       function handleSnapToPosition(
         position: number | string,
-        animationConfigs?: Animated.WithSpringConfig | Animated.WithTimingConfig
+        animationConfigs?: WithSpringConfig | WithTimingConfig
       ) {
         print({
           component: BottomSheet.name,
@@ -893,7 +930,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     );
     const handleClose = useCallback(
       function handleClose(
-        animationConfigs?: Animated.WithSpringConfig | Animated.WithTimingConfig
+        animationConfigs?: WithSpringConfig | WithTimingConfig
       ) {
         print({
           component: BottomSheet.name,
@@ -939,7 +976,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     );
     const handleForceClose = useCallback(
       function handleForceClose(
-        animationConfigs?: Animated.WithSpringConfig | Animated.WithTimingConfig
+        animationConfigs?: WithSpringConfig | WithTimingConfig
       ) {
         print({
           component: BottomSheet.name,
@@ -987,7 +1024,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     );
     const handleExpand = useCallback(
       function handleExpand(
-        animationConfigs?: Animated.WithSpringConfig | Animated.WithTimingConfig
+        animationConfigs?: WithSpringConfig | WithTimingConfig
       ) {
         print({
           component: BottomSheet.name,
@@ -1036,7 +1073,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
     );
     const handleCollapse = useCallback(
       function handleCollapse(
-        animationConfigs?: Animated.WithSpringConfig | Animated.WithTimingConfig
+        animationConfigs?: WithSpringConfig | WithTimingConfig
       ) {
         print({
           component: BottomSheet.name,
@@ -1115,6 +1152,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
         animatedHandleHeight,
         animatedFooterHeight,
         animatedKeyboardHeight,
+        animatedKeyboardHeightInContainer,
         animatedContainerHeight,
         animatedSnapPoints,
         animatedHighestSnapPoint,
@@ -1131,7 +1169,6 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
         failOffsetY: _providedFailOffsetY,
         animateToPosition,
         stopAnimation,
-        getKeyboardHeightInContainer,
         setScrollableRef,
         removeScrollableRef,
       }),
@@ -1149,6 +1186,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
         animatedAnimationState,
         animatedKeyboardState,
         animatedKeyboardHeight,
+        animatedKeyboardHeightInContainer,
         animatedSheetState,
         animatedHighestSnapPoint,
         animatedScrollableState,
@@ -1169,7 +1207,6 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
         _providedActiveOffsetY,
         _providedFailOffsetX,
         _providedFailOffsetY,
-        getKeyboardHeightInContainer,
         setScrollableRef,
         removeScrollableRef,
         animateToPosition,
@@ -1642,9 +1679,7 @@ const BottomSheetComponent = forwardRef<BottomSheet, BottomSheetProps>(
                     key="BottomSheetRootDraggableView"
                     style={contentContainerStyle}
                   >
-                    {typeof children === 'function'
-                      ? (children as Function)()
-                      : children}
+                    {typeof Content === 'function' ? <Content /> : Content}
 
                     {footerComponent && (
                       <BottomSheetFooterContainer