+++ src/components/bottomSheetModal/BottomSheetModal.tsx
@@ -21,6 +21,14 @@ import type { BottomSheetModalProps } from './types';
 
 type BottomSheetModal = BottomSheetModalMethods;
 
+const INITIAL_STATE: {
+  mount: boolean;
+  data: any;
+} = {
+  mount: false,
+  data: undefined,
+};
+
 const BottomSheetModalComponent = forwardRef<
   BottomSheetModal,
   BottomSheetModalProps
@@ -36,17 +44,18 @@ const BottomSheetModalComponent = forwardRef<
     index = 0,
     snapPoints,
     enablePanDownToClose = true,
+    animateOnMount = true,
 
     // callbacks
     onChange: _providedOnChange,
 
     // components
-    children,
+    children: Content,
     ...bottomSheetProps
   } = props;
 
   //#region state
-  const [mount, setMount] = useState(false);
+  const [{ mount, data }, setState] = useState(INITIAL_STATE);
   //#endregion
 
   //#region hooks
@@ -62,7 +71,7 @@ const BottomSheetModalComponent = forwardRef<
 
   //#region refs
   const bottomSheetRef = useRef<BottomSheet>(null);
-  const currentIndexRef = useRef(-1);
+  const currentIndexRef = useRef(!animateOnMount ? index : -1);
   const restoreIndexRef = useRef(-1);
   const minimized = useRef(false);
   const forcedDismissed = useRef(false);
@@ -103,7 +112,7 @@ const BottomSheetModalComponent = forwardRef<
 
       // unmount the node, if sheet is still mounted
       if (_mounted) {
-        setMount(false);
+        setState(INITIAL_STATE);
       }
 
       // fire `onDismiss` callback
@@ -161,9 +170,12 @@ const BottomSheetModalComponent = forwardRef<
 
   //#region bottom sheet modal methods
   const handlePresent = useCallback(
-    function handlePresent() {
+    function handlePresent(_data?: any) {
       requestAnimationFrame(() => {
-        setMount(true);
+        setState({
+          mount: true,
+          data: _data,
+        });
         mountSheet(key, ref, stackBehavior);
 
         print({
@@ -348,7 +360,7 @@ const BottomSheetModalComponent = forwardRef<
   //#endregion
 
   // render
-  // console.log('BottomSheetModal', index, snapPoints)
+  // console.log('BottomSheetModal', index, mount, data);
   return mount ? (
     <Portal
       key={key}
@@ -364,11 +376,14 @@ const BottomSheetModalComponent = forwardRef<
         index={index}
         snapPoints={snapPoints}
         enablePanDownToClose={enablePanDownToClose}
+        animateOnMount={animateOnMount}
         containerHeight={containerHeight}
         containerOffset={containerOffset}
         onChange={handleBottomSheetOnChange}
         onClose={handleBottomSheetOnClose}
-        children={children}
+        children={
+          typeof Content === 'function' ? <Content data={data} /> : Content
+        }
         $modal={true}
       />
     </Portal>