+++ src/components/bottomSheetModal/types.d.ts
@@ -1,3 +1,4 @@
+import type React from 'react';
 import type { BottomSheetProps } from '../bottomSheet';
 import type { MODAL_STACK_BEHAVIOR } from '../../constants';
 
@@ -10,10 +11,7 @@ export interface BottomSheetModalPrivateMethods {
 export type BottomSheetModalStackBehavior = keyof typeof MODAL_STACK_BEHAVIOR;
 
 export interface BottomSheetModalProps
-  extends Omit<
-    BottomSheetProps,
-    'animateOnMount' | 'containerHeight' | 'onClose'
-  > {
+  extends Omit<BottomSheetProps, 'containerHeight' | 'onClose'> {
   /**
    * Modal name to help identify the modal for later on.
    * @type string
@@ -43,4 +41,13 @@ export interface BottomSheetModalProps
    * @type () => void;
    */
   onDismiss?: () => void;
+
+  /**
+   * A scrollable node or normal view.
+   * @type React.ReactNode[] | React.ReactNode
+   */
+  children:
+    | (({ data: any }?) => React.ReactNode)
+    | React.ReactNode[]
+    | React.ReactNode;
 }