+++ src/hooks/useBottomSheetModalInternal.ts
@@ -1,12 +1,25 @@
 import { useContext } from 'react';
-import { BottomSheetModalInternalContext } from '../contexts';
+import {
+  BottomSheetModalInternalContext,
+  BottomSheetModalInternalContextType,
+} from '../contexts';
 
-export const useBottomSheetModalInternal = () => {
+export function useBottomSheetModalInternal(
+  unsafe?: false
+): BottomSheetModalInternalContextType;
+
+export function useBottomSheetModalInternal(
+  unsafe: true
+): BottomSheetModalInternalContextType | null;
+
+export function useBottomSheetModalInternal(
+  unsafe?: boolean
+): BottomSheetModalInternalContextType | null {
   const context = useContext(BottomSheetModalInternalContext);
 
-  if (context === null) {
+  if (unsafe !== true && context === null) {
     throw "'BottomSheetModalInternalContext' cannot be null!";
   }
 
   return context;
-};
+}