+++ CHANGELOG.md
@@ -1,6 +1,78 @@
 # Changelog
 
-## [v4.1.5](https://github.com/gorhom/react-native-bottom-sheet/compare/v4.1.4...v4.1.5)
+## [v4.3.0](https://github.com/gorhom/react-native-bottom-sheet/compare/v4.2.2...v4.3.0)
+
+#### Features
+
+- feat: added data to present modal api ([`#942`](https://github.com/gorhom/react-native-bottom-sheet/pull/942)).
+
+#### Improvements
+
+- refactor: expose animateOnMount for modals ([`#943`](https://github.com/gorhom/react-native-bottom-sheet/pull/943)).
+- refactor: added jest mock file ([`#941`](https://github.com/gorhom/react-native-bottom-sheet/pull/941)).
+
+## [v4.2.2](https://github.com/gorhom/react-native-bottom-sheet/compare/v4.2.1...v4.2.2) - 2022-05-02
+
+#### Fixes
+
+- fix: allowed keyboard height to be recalculated when it changes ([`#931`](https://github.com/gorhom/react-native-bottom-sheet/pull/931)).
+
+## [v4.2.1](https://github.com/gorhom/react-native-bottom-sheet/compare/v4.2.0...v4.2.1) - 2022-04-24
+
+#### Fixes
+
+- fix: updated footer container export name ([a887141](https://github.com/gorhom/react-native-bottom-sheet/commit/a88714153a780395337b84efe00e3d410702c1d9)).
+
+## [v4.2.0](https://github.com/gorhom/react-native-bottom-sheet/compare/v4.1.6...v4.2.0) - 2022-04-24
+
+#### Features
+
+- feat: allow unsafe usage for useBottomSheetInternal & useBottomSheetModalInternal (#740)(by @jembach)
+
+* feat: add useBottomSheetInternal_unsafe & useBottomSheetModalInternal_unsafe to access context
+
+* chore: export hooks useBottomSheetInternal_unsafe and useBottomSheetModalInternal_unsafe
+
+* refactor: combined unsafe hooks with default hooks using an argument
+
+* chore(useBottomSheetInternal): use function overloading instead of interface
+
+* chore(useBottomSheetModalInternal): use function overloading instead of interface
+
+* chore: fixed spacing
+
+Co-authored-by: gorhom &lt;gorhom.dev@gmail.com&gt; ([1bf6139](https://github.com/gorhom/react-native-bottom-sheet/commit/1bf613997cb7a7c8d1fd14f8253701e511a145c7)).
+
+#### Improvements
+
+- chore: fixed types import from reanimated ([831df9c](https://github.com/gorhom/react-native-bottom-sheet/commit/831df9c9e8f25ead974251efcdc384fa1ca00c2e)).
+- chore: fixed types import ([95cb80d](https://github.com/gorhom/react-native-bottom-sheet/commit/95cb80d3331efb12a1b22b904ebdc0155ebcd833)).
+- chore: exported useBottomSheetModalInternal hook ([31eb738](https://github.com/gorhom/react-native-bottom-sheet/commit/31eb73859b46ca325d8960baff9a9ddccb1b89fe)).
+
+## [v4.1.6](https://github.com/gorhom/react-native-bottom-sheet/compare/v4.1.5...v4.1.6) - 2022-04-23
+
+#### Improvements
+
+- chore: updated react native to 0.68 ([b4614bd](https://github.com/gorhom/react-native-bottom-sheet/commit/b4614bdd70a82dc31d9ef148a47533682b67a802)).
+- chore: updated reanimated to 2.8 ([c1e6847](https://github.com/gorhom/react-native-bottom-sheet/commit/c1e6847048c43fb2b678bedfd94ae57502df9765)).
+- chore: added native screens example ([1cf46c0](https://github.com/gorhom/react-native-bottom-sheet/commit/1cf46c08c5561c0320c57e1006b24b70c690a34f)).
+- chore: updated react native portal library ([955b774](https://github.com/gorhom/react-native-bottom-sheet/commit/955b7748932ba5ea81d2406c0acf7b612fecbf0e)).
+- chore: updated portal to 1.0.12 ([0010008](https://github.com/gorhom/react-native-bottom-sheet/commit/0010008906154f9a545f89d5826ea7af48336610)).
+- chore: replaced blacklist with exclusionList (#649)(by @aleppos) ([e3881b3](https://github.com/gorhom/react-native-bottom-sheet/commit/e3881b3149c522102b93c5d2ed2a23003ece4ca2)).
+- chore: export BottomSheetFooterContainer component ([4f63b0d](https://github.com/gorhom/react-native-bottom-sheet/commit/4f63b0d0609160790b420d88478859b91fb8424d)).
+
+#### Fixes
+
+- fix: always update container height to avoid races. (#919)(by @elan)
+
+Updating a shared values is an asynchronous operation, so if two layouts happen in quick succession, we can miss an update. ([3245b23](https://github.com/gorhom/react-native-bottom-sheet/commit/3245b23653a38da2057f28d02f6d2bf1168864d0)).
+- fix: always update handle height to avoid races.(related #919) ([dbf8945](https://github.com/gorhom/react-native-bottom-sheet/commit/dbf894591db8c72c4a0a4a5f1c2986f07ed4b1fb)).
+
+#### Documentations
+
+- docs: updated the readme file ([d951b17](https://github.com/gorhom/react-native-bottom-sheet/commit/d951b17957eb5d2f7f1b40a628ba6d5edd4b5a99)).
+
+## [v4.1.5](https://github.com/gorhom/react-native-bottom-sheet/compare/v4.1.4...v4.1.5) - 2021-12-05
 
 #### Fixes